- Added block entities. - Added active vent block & item. - Fixed non-instanced geometry not being rendered.

+++ src/main/java/org/voxeleers/game/elements/Element.java
@@ -3,12 +3,12 @@ package org.voxeleers.game.elements;
 import org.voxeleers.game.items.ItemType;
 
 public class Element {
-    public float specificHeat;
+    public double specificHeat;
     public int freezingTemp;
     public ItemType iceItemType = null;
     public String name;
 
-    public Element(String name, float specificHeat, int freezingTemp) {
+    public Element(String name, double specificHeat, int freezingTemp) {
         this.name = name;
         this.specificHeat = specificHeat;
         this.freezingTemp = freezingTemp;
